requirements.txt
azure-ai-ml
azure-ai-projects
azure-identity
python-dotenv